Lucene index is now located in the server website path (context path) along with the uploaded images.
The uploaded image folder is created if it does not exist already (will
be wiped by re-publishing unfortunately).

Checkout button is invisible if the user is not logged in.

Search results contain an add to cart button and properly link to item
details.

Cart properly links to item details.

This might actually be the final version finally <:0